change [unsigned] int to size_t where suitable

+++ pyslic3r/_SlicedModel.pxd
@@ -12,18 +12,18 @@ cdef class SlicedModel:
   cdef  s3.SLICEDMODEL *thisptr
   cdef  cnp.ndarray zvalues
   cdef  bool slicesAreOrdered     (self)
-  cdef  void _removeLayers        (self,                                         unsigned int init, unsigned int end) nogil
-  cdef  void _removeExPolygons    (self, unsigned int nlayer,                    unsigned int init, unsigned int end) nogil
-  cdef  void _removeHoles         (self, unsigned int nlayer, unsigned int nexp, unsigned int init, unsigned int end) nogil
-  cpdef unsigned int numLayers    (self)
-  cpdef unsigned int numExPolygons(self, unsigned int nlayer)
-  cpdef unsigned int numHoles     (self, unsigned int nlayer, unsigned int nexp)
+  cdef  void _removeLayers        (self,                                         size_t init, size_t end) nogil
+  cdef  void _removeExPolygons    (self, size_t nlayer,                    size_t init, size_t end) nogil
+  cdef  void _removeHoles         (self, size_t nlayer, size_t nexp, size_t init, size_t end) nogil
+  cpdef size_t numLayers          (self)
+  cpdef size_t numExPolygons      (self, size_t nlayer)
+  cpdef size_t numHoles           (self, size_t nlayer, size_t nexp)
   cpdef object  toLayerList       (self,                                         bool asInteger=?, bool asView=?, rang=?)
-  cpdef object  toExPolygonList   (self, unsigned int nlayer,                    bool asInteger=?, bool asView=?, rang=?)
-  cdef  object _toExPolygonList   (self, unsigned int nlayer,                    bool asInteger=?, bool asView=?, rang=?)
-  cpdef object  toHoleList        (self, unsigned int nlayer, unsigned int nexp, bool asInteger=?, bool asView=?, rang=?)
-  cdef  object _toHoleList        (self, unsigned int nlayer, unsigned int nexp, bool asInteger=?, bool asView=?, rang=?)
+  cpdef object  toExPolygonList   (self, size_t nlayer,                    bool asInteger=?, bool asView=?, rang=?)
+  cdef  object _toExPolygonList   (self, size_t nlayer,                    bool asInteger=?, bool asView=?, rang=?)
+  cpdef object  toHoleList        (self, size_t nlayer, size_t nexp, bool asInteger=?, bool asView=?, rang=?)
+  cdef  object _toHoleList        (self, size_t nlayer, size_t nexp, bool asInteger=?, bool asView=?, rang=?)
   cdef  cnp.ndarray Polygon2array (self, s3.Polygon *pol,                        bool asInteger=?, bool asView=?)
-  cdef  c.Paths* _layerToClipperPaths        (self, unsigned int nlayer, c.Paths    *output) nogil
-  cdef  void     _setLayerFromClipperPaths   (self, unsigned int nlayer, c.Paths    *inputs) nogil
-  cdef  void     _setLayerFromClipperPolyTree(self, unsigned int nlayer, c.PolyTree *inputs) nogil
+  cdef  c.Paths* _layerToClipperPaths        (self, size_t nlayer, c.Paths    *output) nogil
+  cdef  void     _setLayerFromClipperPaths   (self, size_t nlayer, c.Paths    *inputs) nogil
+  cdef  void     _setLayerFromClipperPolyTree(self, size_t nlayer, c.PolyTree *inputs) nogil
